LPS-54824 Make CustomAttributesDisplay package consistent across modules
